+++ packages/components/src/components/Button/Button.tsx
@@ -1,11 +1,11 @@
-import React, { FC, PropsWithChildren } from "react";
+import React, { PropsWithChildren } from "react";
 import styles from "./Button.module.scss";
 import * as Aria from "react-aria-components";
 import clsx from "clsx";
 import {
+  ClearPropsContext,
   PropsContext,
   PropsContextProvider,
-  useProps,
 } from "@/lib/propsContext";
 import {
   IconFailed,
@@ -14,9 +14,16 @@ import {
 } from "@/components/Icon/components/icons";
 import { Wrap } from "@/components/Wrap";
 import { Text } from "@/components/Text";
+import {
+  flowComponent,
+  FlowComponentProps,
+} from "@/lib/componentFactory/flowComponent";
+import locales from "./locales/*.locale.json";
+import { useLocalizedStringFormatter } from "react-aria";
 
 export interface ButtonProps
-  extends PropsWithChildren<Omit<Aria.ButtonProps, "style">> {
+  extends PropsWithChildren<Omit<Aria.ButtonProps, "style">>,
+    FlowComponentProps {
   /** @default "primary" */
   variant?: "primary" | "accent" | "secondary" | "danger";
   /** @default "solid" */
@@ -29,7 +36,24 @@ export interface ButtonProps
   isFailed?: boolean;
 }
 
-export const Button: FC<ButtonProps> = (props) => {
+const disablePendingProps = (props: ButtonProps) => {
+  if (props.isPending || props.isSucceeded || props.isFailed) {
+    props = { ...props };
+    props.onPress = undefined;
+    props.onPressStart = undefined;
+    props.onPressEnd = undefined;
+    props.onPressChange = undefined;
+    props.onPressUp = undefined;
+    props.onKeyDown = undefined;
+    props.onKeyUp = undefined;
+  }
+
+  return props;
+};
+
+export const Button = flowComponent("Button", (props) => {
+  props = disablePendingProps(props);
+
   const {
     variant = "primary",
     style = "solid",
@@ -40,8 +64,9 @@ export const Button: FC<ButtonProps> = (props) => {
     isDisabled,
     isSucceeded,
     isFailed,
+    "aria-label": ariaLabel,
     ...restProps
-  } = useProps("Button", props);
+  } = props;
 
   const rootClassName = clsx(
     styles.button,
@@ -65,6 +90,17 @@ export const Button: FC<ButtonProps> = (props) => {
     },
   };
 
+  const stringFormatter = useLocalizedStringFormatter(locales);
+
+  const stateLabel =
+    isSucceeded || isFailed || isPending
+      ? stringFormatter.format(
+          `button.${
+            isSucceeded ? "isSucceeded" : isFailed ? "isFailed" : "isPending"
+          }`,
+        )
+      : undefined;
+
   const StateIconComponent = isSucceeded
     ? IconSucceeded
     : isFailed
@@ -80,23 +116,28 @@ export const Button: FC<ButtonProps> = (props) => {
   const isStringContent = typeof children === "string";
 
   return (
-    <Aria.Button
-      className={rootClassName}
-      isDisabled={isDisabled || isPending || isSucceeded || isFailed}
-      {...restProps}
-    >
-      <PropsContextProvider props={propsContext}>
+    <ClearPropsContext>
+      <Aria.Button
+        className={rootClassName}
+        isDisabled={isDisabled}
+        aria-label={stateLabel ?? ariaLabel}
+        {...restProps}
+      >
         <Wrap if={stateIcon}>
           <span className={styles.content}>
             <Wrap if={isStringContent}>
-              <Text>{children}</Text>
+              <Text>
+                <PropsContextProvider props={propsContext}>
+                  {children}
+                </PropsContextProvider>
+              </Text>
             </Wrap>
           </span>
         </Wrap>
-      </PropsContextProvider>
-      {stateIcon}
-    </Aria.Button>
+        {stateIcon}
+      </Aria.Button>
+    </ClearPropsContext>
   );
-};
+});
 
 export default Button;