+++ packages/components/src/components/Content/Content.tsx
@@ -1,19 +1,26 @@
-import React, { FC, PropsWithChildren } from "react";
-import { ClearPropsContext, useProps } from "@/lib/propsContext";
+import React, { PropsWithChildren } from "react";
+import { ClearPropsContext } from "@/lib/propsContext";
 import { PropsWithElementType } from "@/lib/types/props";
+import {
+  flowComponent,
+  FlowComponentProps,
+} from "@/lib/componentFactory/flowComponent";
 
-export interface ContentProps extends PropsWithChildren, PropsWithElementType {}
+export interface ContentProps
+  extends PropsWithChildren,
+    PropsWithElementType,
+    FlowComponentProps {}
 
-export const Content: FC<ContentProps> = (props) => {
-  const { children, elementType = "div", ...rest } = useProps("Content", props);
+export const Content = flowComponent("Content", (props) => {
+  const { children, elementType = "div", ...rest } = props;
 
   const Element = elementType;
 
   return (
-    <Element {...rest}>
-      <ClearPropsContext>{children}</ClearPropsContext>
-    </Element>
+    <ClearPropsContext>
+      <Element {...rest}>{children}</Element>
+    </ClearPropsContext>
   );
-};
+});
 
 export default Content;