+++ packages/components/src/components/Initials/Initials.tsx
@@ -1,17 +1,23 @@
-import React, { FC, PropsWithChildren } from "react";
+import React, { PropsWithChildren } from "react";
 import { getInitialsFromString } from "./lib/getInitialsFromString";
 import styles from "./Initials.module.scss";
 import clsx from "clsx";
-import { useProps } from "@/lib/propsContext";
+import { ClearPropsContext } from "@/lib/propsContext";
+import { onlyText } from "react-children-utilities";
+import {
+  flowComponent,
+  FlowComponentProps,
+} from "@/lib/componentFactory/flowComponent";
 
-export interface InitialsProps extends PropsWithChildren<{ children: string }> {
+export interface InitialsProps extends PropsWithChildren, FlowComponentProps {
   className?: string;
 }
 
-export const Initials: FC<InitialsProps> = (props) => {
-  const { children, className } = useProps("Initials", props);
+export const Initials = flowComponent("Initials", (props) => {
+  const { children, className } = props;
 
-  const initials = getInitialsFromString(children);
+  const textContent = onlyText(children);
+  const initials = getInitialsFromString(textContent);
 
   const rootClassName = clsx(styles.initials, className);
 
@@ -20,10 +26,12 @@ export const Initials: FC<InitialsProps> = (props) => {
   ));
 
   return (
-    <div aria-label={children} className={rootClassName}>
-      {initialsElements}
-    </div>
+    <ClearPropsContext>
+      <div aria-label={textContent} className={rootClassName}>
+        {initialsElements}
+      </div>
+    </ClearPropsContext>
   );
-};
+});
 
 export default Initials;