+++ packages/components/src/components/LabeledValue/LabeledValue.tsx
@@ -1,7 +1,11 @@
 import React, { FC, PropsWithChildren } from "react";
 import styles from "./LabeledValue.module.scss";
 import clsx from "clsx";
-import { PropsContext, PropsContextProvider } from "@/lib/propsContext";
+import {
+  ClearPropsContext,
+  PropsContext,
+  PropsContextProvider,
+} from "@/lib/propsContext";
 
 export interface LabeledValueProps extends PropsWithChildren {
   className?: string;
@@ -29,11 +33,13 @@ export const LabeledValue: FC<LabeledValueProps> = (props) => {
   };
 
   return (
-    <div className={rootClassName}>
-      <PropsContextProvider props={propsContext}>
-        {children}
-      </PropsContextProvider>
-    </div>
+    <ClearPropsContext>
+      <div className={rootClassName}>
+        <PropsContextProvider props={propsContext}>
+          {children}
+        </PropsContextProvider>
+      </div>
+    </ClearPropsContext>
   );
 };