+++ packages/components/src/components/List/components/OptionsButton/index.tsx
@@ -1,0 +1,3 @@
+import { OptionsButton } from "./OptionsButton";
+export { OptionsButton } from "./OptionsButton";
+export default OptionsButton;