+++ packages/components/src/components/StatusIcon/StatusIcon.tsx
@@ -9,6 +9,7 @@ import locales from "./locales/*.locale.json";
 import { useLocalizedStringFormatter } from "react-aria";
 import { PropsWithStatus, Status } from "@/lib/types/props";
 import { IconProps } from "@/components/Icon";
+import { ClearPropsContext } from "@/lib/propsContext";
 
 export interface StatusIconProps extends PropsWithStatus, IconProps {}
 
@@ -28,7 +29,11 @@ export const StatusIcon: FC<StatusIconProps> = (props) => {
 
   const Icon = icons[status];
 
-  return <Icon aria-label={ariaLabel} {...rest} />;
+  return (
+    <ClearPropsContext>
+      <Icon aria-label={ariaLabel} {...rest} />
+    </ClearPropsContext>
+  );
 };
 
 export default StatusIcon;