+++ packages/design-tokens/CHANGELOG.md
@@ -3,6 +3,60 @@
 All notable changes to this project will be documented in this file. See
 [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.1.0-alpha.55](https://github.com/mittwald/flow/compare/0.1.0-alpha.54...0.1.0-alpha.55) (2024-03-26)
+
+### Features
+
+- add context menu component
+  ([0fb5ec1](https://github.com/mittwald/flow/commit/0fb5ec12e3dd097fb554331c87cf3dc80f70b3c8))
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [0.1.0-alpha.53](https://github.com/mittwald/flow/compare/0.1.0-alpha.52...0.1.0-alpha.53) (2024-03-25)
+
+**Note:** Version bump only for package @mittwald/flow-design-tokens
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [0.1.0-alpha.50](https://github.com/mittwald/flow/compare/0.1.0-alpha.49...0.1.0-alpha.50) (2024-03-22)
+
+### Features
+
+- add Modal component
+  ([d9a47ea](https://github.com/mittwald/flow/commit/d9a47eae58db9c7a8b1b6868da7246e1eded2a16))
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [0.1.0-alpha.43](https://github.com/mittwald/flow/compare/0.1.0-alpha.42...0.1.0-alpha.43) (2024-03-19)
+
+**Note:** Version bump only for package @mittwald/flow-design-tokens
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [0.1.0-alpha.41](https://github.com/mittwald/flow/compare/0.1.0-alpha.40...0.1.0-alpha.41) (2024-03-19)
+
+### Features
+
+- support Switch in Sections
+  ([2becdb5](https://github.com/mittwald/flow/commit/2becdb5a891a7ef263aad72027d65c8e24ae67d9))
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
 # [0.1.0-alpha.37](https://github.com/mittwald/flow/compare/0.1.0-alpha.36...0.1.0-alpha.37) (2024-03-18)
 
 **Note:** Version bump only for package @mittwald/flow-design-tokens